[Refactor] path_finder.py 파일 분리 및 dead code 제거
